Removing default mobile first breakpoints
Given the design I've been provided with is desktop size, that is where I have built to so far. I will now build DOWN to mobile. This Bones theme was initially designed to build mobile first, but is easy enough to change.

I believe Mobile-First ideology is bad practise: the name "Mobile-First" implies mobile will be better than desktop. Websites should be built to work perfectly on all screen sizes, so (in the majority of cases, including this one) it doesn't matter what size is built first or whether the media queries are min or max.